roachtest/mixedversion: move helper functions to its own file
This makes it easier for test authors to see all helper functions
available without having to browse through internal framework code.

Epic: CRDB-19321

Release note: None
