+++ pkg/cmd/roachtest/roachtestutil/mixedversion/BUILD.bazel
@@ -3,6 +3,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "mixedversion",
     srcs = [
+        "helper.go",
         "mixedversion.go",
         "planner.go",
         "runner.go",
